move into threading integration
