--- sentry_sdk/integrations/concurrent.py
@@ -1,50 +1,0 @@
-from functools import wraps
-
-from concurrent.futures import ThreadPoolExecutor, Future
-
-import sentry_sdk
-from sentry_sdk.integrations import Integration
-from sentry_sdk.scope import use_isolation_scope, use_scope
-
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from typing import Any, Callable, TypeVar
-
-    T = TypeVar("T", bound=Any)
-
-
-class ConcurrentIntegration(Integration):
-    identifier = "concurrent"
-
-    @staticmethod
-    def setup_once():
-        # type: () -> None
-        ThreadPoolExecutor.submit = _wrap_submit_call(ThreadPoolExecutor.submit)  # type: ignore
-
-
-def _wrap_submit_call(func):
-    # type: (Callable[..., Future[T]]) -> Callable[..., Future[T]]
-    """
-    Wrap submit call to propagate scopes on task submission.
-    """
-
-    @wraps(func)
-    def sentry_submit(self, fn, *args, **kwargs):
-        # type: (ThreadPoolExecutor, Callable[..., T], *Any, **Any) -> Future[T]
-        integration = sentry_sdk.get_client().get_integration(ConcurrentIntegration)
-        if integration is None:
-            return func(self, fn, *args, **kwargs)
-
-        isolation_scope = sentry_sdk.get_isolation_scope().fork()
-        current_scope = sentry_sdk.get_current_scope().fork()
-
-        def wrapped_fn(*args, **kwargs):
-            # type: (*Any, **Any) -> Any
-            with use_isolation_scope(isolation_scope):
-                with use_scope(current_scope):
-                    return fn(*args, **kwargs)
-
-        return func(self, wrapped_fn, *args, **kwargs)
-
-    return sentry_submit
